# contracts/api/marketplace-central.openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace Central API
  version: 2026-07-13
paths:
...
  /config/active-source:
    get:
      summary: Get the tenant's active ERP/catalog source
      operationId: getActiveSource
      tags: [tenant-config]
      responses:
        "200":
          description: Tenant active-source configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveSourceConfig'
        "400":
          description: Tenant has no active-source row (fail-closed, never a silent default)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveSourceError'
    put:
      summary: Set the tenant's active ERP/catalog source
      operationId: setActiveSource
      tags: [tenant-config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetActiveSourceRequest'
      responses:
        "200":
          description: Stored tenant active-source configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveSourceConfig'
        "400":
          description: Invalid active_source or malformed body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveSourceError'
  /config/sellable-assortment:
    get:
      summary: Get the tenant's sellable-assortment filters
      operationId: getSellableAssortment
      tags: [tenant-config]
      responses:
        "200":
          description: Tenant sellable-assortment configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellableAssortmentConfig'
        "400":
          description: Tenant has no active ERP source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellableAssortmentError'
        "500":
          description: Tenant sellable-assortment configuration could not be read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellableAssortmentError'
    put:
      summary: Set the tenant's sellable-assortment filters
      operationId: setSellableAssortment
      tags: [tenant-config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSellableAssortmentRequest'
      responses:
        "200":
          description: Stored tenant sellable-assortment configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellableAssortmentConfig'
        "400":
          description: Invalid sellable-assortment body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellableAssortmentError'
        "500":
          description: Tenant sellable-assortment configuration could not be stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellableAssortmentError'
components:
...
  schemas:
...
    ActiveSourceConfig:
      type: object
      description: A tenant's active ERP/catalog source selection, as stored.
      required: [active_source, source_kind, set_at]
      properties:
        active_source:
          type: string
          enum: [sankhya, xlsx, catalogo_cliente]
        source_kind:
          type: string
          description: Server-derived from active_source (tenant_config.DefaultKind) — never client-supplied.
          enum: [upload_snapshot, live_read_through]
        set_at:
          type: string
          format: date-time
        set_by:
          type: string
          nullable: true
    SetActiveSourceRequest:
      type: object
      description: >-
        Selects the tenant's active source. source_kind is intentionally absent:
        it is always derived server-side from active_source and any client-supplied
        value is ignored.
      required: [active_source]
      properties:
        active_source:
          type: string
          enum: [sankhya, xlsx, catalogo_cliente]
        set_by:
          type: string
          nullable: true
    ActiveSourceError:
      type: object
      description: Flat active-source error body.
      required: [error]
      properties:
        error:
          type: string
          enum: [unknown_erp_source, invalid_active_source, invalid_body, internal_error]
        detail:
          type: string
    SellableAssortmentConfig:
      type: object
      description: Tenant's configured filters for the sellable catalog assortment.
      required: [only_revenda, only_em_estoque, only_ecommerce_eligible]
      properties:
        only_revenda:
          type: boolean
        only_em_estoque:
          type: boolean
        only_ecommerce_eligible:
          type: boolean
    SetSellableAssortmentRequest:
      type: object
      description: Replacement values for the tenant's sellable-assortment filters.
      required: [only_revenda, only_em_estoque, only_ecommerce_eligible]
      properties:
        only_revenda:
          type: boolean
        only_em_estoque:
          type: boolean
        only_ecommerce_eligible:
          type: boolean
    SellableAssortmentError:
      type: object
      description: >-
        Flat sellable-assortment error body, mirroring ActiveSourceError. The
        tenant_config transport writes map[string]string{"error": code} plus an
        optional "detail", so the nested ErrorResponse shape is not what this
        surface emits.
      required: [error]
      properties:
        error:
          type: string
          enum: [unknown_erp_source, invalid_body, internal_error]
        detail:
          type: string
...